backend/requirements.txt
# Core Web Framework
fastapi==0.115.5
uvicorn[standard]==0.32.1
starlette==0.41.3

# Data Validation & Settings
pydantic==2.10.3
pydantic-core==2.27.2
pydantic-settings==2.6.1

# Configuration Management
python-dotenv==1.0.1

# Graph Processing
networkx==3.4.2

# Neo4j Database Driver (for production)
neo4j==5.16.0

# LLM Integration (Google Gemini)
google-generativeai==0.8.3
google-ai-generativelanguage==0.6.10

# Data Processing
pandas==2.2.3

# HTTP Multipart & CORS
python-multipart==0.0.20
fastapi-cors==0.0.6

# File Operations
aiofiles==24.1.0
